chore(alerts): reuse shared breakdown-label humanizer
The new alerts formatting.py added `humanize_breakdown_label`, a byte-for-byte
duplicate of `replace_breakdown_labels` in ee/hogai — both wrapping the same
`.replace()` over the sentinel constants from `insights/utils/breakdowns.py`.

Hoist a single `humanize_breakdown_label` into that breakdowns module, next to
the sentinel constants and display strings it maps, and point both the alerts
trends extractor and the hogai trends formatter at it. Drop the now-unused
constant imports from both former call sites.

Consolidate the two redundant unit tests (one in alerts, one in ee/hogai) into
one canonical parameterized test alongside the function in test_breakdowns.py.

No behavior change: the humanizer logic is identical, only relocated.

Generated-By: PostHog Code
Task-Id: 2c5b7b95-2d5b-4f70-8d76-dc1ae7b7c237
