+++ posthog/hogql_queries/insights/utils/test/test_breakdowns.py
@@ -6,6 +6,7 @@ from posthog.hogql_queries.insights.utils.breakdowns import (
     has_breakdown_filter,
     has_multi_breakdown,
     has_single_breakdown,
+    humanize_breakdown_label,
 )
 
 
@@ -44,3 +45,25 @@ def test_breakdown_presence_helpers(
     assert has_breakdown_filter(breakdown_filter) is expected_has_breakdown
     assert has_single_breakdown(breakdown_filter) is expected_has_single
     assert has_multi_breakdown(breakdown_filter) is expected_has_multi
+
+
+@pytest.mark.parametrize(
+    "label, expected",
+    [
+        ("$$_posthog_breakdown_other_$$", "Other (i.e. all remaining values)"),
+        ("$$_posthog_breakdown_null_$$", "None (i.e. no value)"),
+        # both sentinels in one label are each replaced
+        (
+            "$$_posthog_breakdown_other_$$ and $$_posthog_breakdown_null_$$",
+            "Other (i.e. all remaining values) and None (i.e. no value)",
+        ),
+        # compound label shapes survive: action-prefixed and "::"-joined multi-breakdown
+        ("signed_up - $$_posthog_breakdown_other_$$", "signed_up - Other (i.e. all remaining values)"),
+        ("$$_posthog_breakdown_other_$$::US", "Other (i.e. all remaining values)::US"),
+        ("Chrome", "Chrome"),
+        # a normal label containing " - " must pass through untouched (no fragile splitting)
+        ("Signed up - paid", "Signed up - paid"),
+    ],
+)
+def test_humanize_breakdown_label(label: str, expected: str) -> None:
+    assert humanize_breakdown_label(label) == expected
