+++ products/alerts/backend/evaluation/trends.py
@@ -15,6 +15,7 @@ from posthog.api.services.query import ExecutionMode
 from posthog.caching.calculate_results import calculate_for_query_based_insight
 from posthog.caching.fetch_from_cache import InsightResult
 from posthog.event_usage import EventSource
+from posthog.hogql_queries.insights.utils.breakdowns import humanize_breakdown_label
 
 # These helpers also back the anomaly detector, so they remain in the tasks module for now.
 from posthog.tasks.alerts.trends import (
@@ -31,7 +32,7 @@ from products.alerts.backend.evaluation.contract import (
     lookback_intervals_for,
     zero_sentinel_series,
 )
-from products.alerts.backend.evaluation.formatting import humanize_breakdown_label, make_trends_value_formatter
+from products.alerts.backend.evaluation.formatting import make_trends_value_formatter
 from products.alerts.backend.models.alert import AlertConfiguration
 from products.product_analytics.backend.models.insight import Insight
